4.21 done

+++ controllers/blogs.js
@@ -1,4 +1,5 @@
 const blogsRouter = require('express').Router()
+const blog = require('../models/blog')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
@@ -42,10 +43,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
-  if (!deletedBlog) {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
     return response.status(404).send({ error: 'Blog not found' })
   }
+  if (blog.user.toString() !== decodedToken.id.toString()){
+    return response.status(403).send({ error: 'Only the creator can delete this blog' })
+  }
+  await Blog.findByIdAndDelete(request.params.id)
+
   response.status(204).end()
 })
